chore 📦: Remove boilerplate example and update .gitignore
This commit removes the `boilerplate-example.py` file from the `docs` directory as it is no longer needed, and adds a `.gitkeep` file to the `output` directory to prevent its removal by Git. Additionally, the `.gitignore` file has been updated to exclude the `output` directory while keeping the `.gitkeep` file intact. These changes ensure that the project remains organized and that the output directory is not accidentally deleted from the repository.